Rename the directory to be generic
