+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/ai-image/usage-counter.tsx
