+++ projects/plugins/jetpack/extensions/shared/external-media/sources/jetpack-ai-featured-image.js
@@ -1,6 +1,6 @@
 import FeaturedImage, {
 	FEATURED_IMAGE_PLACEMENT_MEDIA_SOURCE_DROPDOWN,
-} from '../../../plugins/ai-assistant-plugin/components/featured-image';
+} from '../../../plugins/ai-assistant-plugin/components/ai-image';
 
 function JetpackAIFeaturedImage( { onClose = () => {} } ) {
 	return (
